requirements.txt
Flask==3.1.1
graphene==3.4.3
Flask-Cors==6.0.0
python-dotenv==1.1.0 
requests==2.32.3
pytest==8.3.5